# Ref String() and greatly improved builtin lookup cost