--- a/v1/topdown/eval.go
+++ b/v1/topdown/eval.go
@@ -171,16 +171,17 @@ func (e *eval) string(s *strings.Builder) {
 func (e *eval) builtinFunc(name string) (*ast.Builtin, BuiltinFunc, bool) {
 	decl, ok := ast.BuiltinMap[name]
 	if ok {
-		f, ok := builtinFunctions[name]
-		if ok {
+		if f, ok := builtinFunctions[name]; ok {
 			return decl, f, true
 		}
-	} else {
-		bi, ok := e.builtins[name]
-		if ok {
-			return bi.Decl, bi.Func, true
+		if bi, ok := e.builtins[name]; ok {
+			return decl, bi.Func, true
 		}
 	}
+	if bi, ok := e.builtins[name]; ok {
+		return bi.Decl, bi.Func, true
+	}
+
 	return nil, nil, false
 }
 
@@ -951,7 +952,7 @@ func (e *eval) evalCall(terms []*ast.Term, iter unifyIterator) error {
 	var bctx *BuiltinContext
 
 	// Creating a BuiltinContext is expensive, so only do it if the builtin depends on it.
-	if bi.NeedsBuiltInContext() {
+	if !bi.CanSkipBctx {
 		var parentID uint64
 		if e.parent != nil {
 			parentID = e.parent.queryID