@anderseknert 2025-10-04
The big difference from before is of course the "dot builtin" case, which no longer allocates, and completes an order of magnitude faster than before. The "with escape" improvement is pretty nice too.


Main
```
BenchmarkRefString/dot_builtin-16         	 7532424	       157.7 ns/op	      32 B/op	       1 allocs/op
BenchmarkRefString/really_long-16         	 1779793	       668.3 ns/op	     112 B/op	       1 allocs/op
BenchmarkRefString/scalars_ref-16         	 3383953	       355.1 ns/op	      96 B/op	       1 allocs/op
BenchmarkRefString/simple_ref-16          	 6937478	       170.7 ns/op	      32 B/op	       1 allocs/op
BenchmarkRefString/var_term-16            	486741614	         2.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkRefString/with_escape-16         	 6436400	       185.1 ns/op	      56 B/op	       3 allocs/op
```

Now
```
BenchmarkRefString/dot_builtin-16         	74980562	        15.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkRefString/really_long-16         	 1797901	       665.4 ns/op	      80 B/op	       1 allocs/op
BenchmarkRefString/scalars_ref-16         	 3481281	       343.5 ns/op	      64 B/op	       1 allocs/op
BenchmarkRefString/simple_ref-16          	 6643372	       179.7 ns/op	      64 B/op	       1 allocs/op
BenchmarkRefString/var_term-16            	423616689	         2.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkRefString/with_escape-16         	 6938409	       172.5 ns/op	      64 B/op	       1 allocs/op
```